Update: Separate rate limit from routes.js to rateLimit.js
